flask/requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysql-connector-python==8.0.11
protobuf==3.6.0
requests==2.19.1
six==1.11.0
urllib3==1.23
uWSGI==2.0.17.1
Werkzeug==0.14.1
